Removed all references to pool.lock outside pool.py
I think what was happening and causing lockups is
pool.py was being acquired by lp.py in some sort of circular manner.
This was locking out the wsgi thread. Which was causing all sorts of disasters

In the long term I want to remove all of the locks but for now I am removing
pool.lock access outside the pool.py module.